Moved out json logic from mosto.js to json-driver.js
